% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_ex_to_nb}
\alias{convert_ex_to_nb}
\title{Convert Exercise To Rmarkdown Notebook}
\usage{
convert_ex_to_nb(
  ps_ex_path,
  ps_nb_out_dir,
  ps_nb_deploy_dir,
  pb_force = FALSE,
  pb_debug = FALSE,
  pobj_rtt_logger = NULL
)
}
\arguments{
\item{ps_ex_path}{path to the Rmarkdown source file}

\item{ps_nb_out_dir}{output directory for Rmd notebook}

\item{ps_nb_deploy_dir}{deployment directory of rendered nb}

\item{pb_force}{force output even if output file already exists}

\item{pb_debug}{debugging flag}

\item{pobj_rtt_logger}{log4r logger object}
}
\description{
Based on the input of a Rmarkdown source file containing exercise problems
and solutions, a Rmarkdown notebook is created that contains only the
exercise problems.
}
\details{
This function assumes that the solutions are marked with surrounding R-code
junks with the tags comment-start and comment-end.
}
\examples{
\dontrun{
convert_ex_to_nb(ps_ex_path = 'ex/asm_ex02.Rmd',
                 ps_nb_out_dir = 'nb',
                 ps_nb_deploy_dir = 'docs/nb')
}

}
