% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_exam.R
\name{create_exam}
\alias{create_exam}
\title{Create New Exam Document}
\usage{
create_exam(ps_exam_path, pl_data = NULL, pb_edit = FALSE, pb_force = FALSE)
}
\arguments{
\item{ps_exam_path}{path to new exam document}

\item{pl_data}{list of data for replacement of placeholders}

\item{pb_edit}{flag whether new document should directly be opened}

\item{pb_force}{flag whether old version should be overwritten}
}
\description{
A new exam document is created based on a predefined template that comes
with package \code{rteachtools}.
}
