% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ex.R
\name{create_exercise}
\alias{create_exercise}
\title{Create Course Exercise From Template}
\usage{
create_exercise(ps_ex_path, pn_nr_problem, pl_data)
}
\arguments{
\item{ps_ex_path}{path to output file where exercise should be written to}

\item{pn_nr_problem}{number of problems in exercise}

\item{pl_data}{list of replacement data for replacing placeholder}
}
\description{
The template rttexercise in the package rteachtools is used to create a new
exercise. The use of this template-based creation approach assures that the
structure of the source document is consistent with the requirements of the
deployment tools.
}
\details{
The function \code{rmarkdown::draft()} is used to convert the template into
the target document. Further, the placeholders in the template are replaced
by specific values using the function \code{whisker::whisker_render()}.
}
\examples{
\dontrun{
create_exercise(ps_ex_path    = "ex/gel_ex01",
                pn_nr_problem = 2,
                pl_data       = list(course_name = "Genetic Evaluation",
                                     exercise_count = 1,
                                     creation_date = format(Sys.Date(), "\%Y-\%m-\%d"),
                                     author = "Peter von Rohr"))
}
}
