% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slides.R
\name{create_slides}
\alias{create_slides}
\title{Create New Beamer Slides}
\usage{
create_slides(
  ps_sl_path,
  ps_template = "cbeamerdown",
  ps_package = "rteachtools",
  pb_edit = FALSE,
  pl_data
)
}
\arguments{
\item{ps_sl_path}{path of the new rmarkdown document}

\item{ps_template}{template name to be used}

\item{ps_package}{package that defines the template}

\item{pb_edit}{should newly created document be directly opened for editing}

\item{pl_data}{list of place-holder-value pairs for replacement.}
}
\description{
A new document in the format of beamer slides are created. The document
is based on a template that can be chosen using a template name and the
package where the template is defined. The default is to use the 'cbeamerdown'
template of package 'rteachtools'.
}
\details{
The rmarkdown document is created using the function \code{rmarkdown::draft()}.
Then some placeholders are replaced using the function \code{whisker::whisker.render()}.
}
\examples{
\dontrun{
create_slides(ps_sl_path = "sl/sl_l01_intro/sl_l01_intro.Rmd",
              pl_data = list(title = "Introduction to Livestock Breeding and Genomics",
                             author = "Peter von Rohr"))
}
}
