% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_ex}
\alias{deploy_ex}
\title{Deploy Exercise}
\usage{
deploy_ex(
  ps_ex_path,
  ps_ex_out_dir,
  ps_sol_out_dir,
  pb_debug = FALSE,
  pobj_rtt_logger = NULL
)
}
\arguments{
\item{ps_ex_path}{path to the source RMarkdown file}

\item{ps_ex_out_dir}{directory for exercise output}

\item{ps_sol_out_dir}{directory for the solution output}

\item{pb_debug}{flag to determine debugging status}

\item{pobj_rtt_logger}{log4j logger object}
}
\description{
An parametrized Rmd source file is rendered two times once to produce a pdf-output
that is used as an excercise. The second pdf-output corresponds to the solution
to the exercises.
}
\details{
The call to \code{rmarkdown::render()} takes the value for the parameters which
are used to produce the exercise or the solution document.
}
\examples{
\dontrun{
deploy_ex(ps_ex_path = 'ex/asm_ex02.Rmd', ps_ex_out_dir = 'docs/ex', ps_sol_out_dir = 'docs/sol')
}

}
