% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_info.R
\name{extract_student_info}
\alias{extract_student_info}
\title{Extract Student Info for R Exercise Platform}
\usage{
extract_student_info(
  ps_bl_xlsx = file.path(here::here(), "students"),
  pvec_col_keep = c("Familienname", "Rufname", "Nummer", "E-Mail"),
  pvec_username = NULL,
  pn_start_port = 10087
)
}
\arguments{
\item{ps_bl_xlsx}{directory or path where input xlsx file can be found}

\item{pvec_col_keep}{columns that are kept from original data}

\item{pn_start_port}{start of mapped port on host}
}
\value{
tibble with relevant student information
}
\description{
Take an xlsx input file downloaded from the student information platform
edoz and extract all information relevant for the R exercise platform.
}
\details{
Usernames are created based on the part of the E-Mail address before the
at sign. By default, the xlsx student input file is searched in the
subdirectory "students" of the current RStudio-project where the path to
this project is determined by the result of \code{here::here()}.
}
\examples{
\dontrun{
s_bl_xlsx <- file.path(here::here(), "students")
extract_student_info(ps_bl_xlsx = s_bl_xlsx)

}
}
