% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_info.R
\name{get_random_student_id}
\alias{get_random_student_id}
\title{Generate Random Student ID Number}
\usage{
get_random_student_id(
  pl_year_limit = list(low = as.integer(format(Sys.Date(), "\%y")) - 4L, up =
    as.integer(format(Sys.Date(), "\%y")) - 2L),
  pl_middle_limit = list(low = 900, up = 999),
  pl_last_limit = list(low = 100, up = 199)
)
}
\arguments{
\item{l_year_limit}{limits for first part}

\item{l_middle_limit}{limits for middle part}

\item{l_last_limit}{limits of last part}
}
\value{
generated student ID
}
\description{
The student ID number consists of three parts. These parts are integer numbers
out of some assumed ranges. From these ranges numbers are sampled from a uniform
distribution
}
\examples{
\dontrun{
get_random_student_id()
}
}
