% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_rtt_logger}
\alias{get_rtt_logger}
\title{Create log4r Logger for package}
\usage{
get_rtt_logger(ps_logfile = "rtt.log", ps_level = "FATAL")
}
\arguments{
\item{ps_logfile}{name of the logfile}

\item{ps_level}{logger level}
}
\value{
rtt_logger
}
\description{
Create log4r Logger for package
}
\examples{
\dontrun{
rtt_logger <- get_rtt_logger()
}
}
