% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_seed.R
\name{get_seed}
\alias{get_seed}
\title{Create a Uniformly Distributed Random Seed}
\usage{
get_seed(pn_min = 1000L, pn_max = 9999L, pn_seed = NULL)
}
\arguments{
\item{pn_min}{lower bound of seed}

\item{pn_max}{upper bound of seed}

\item{pn_seed}{pre-defined seed for debugging}
}
\description{
Random uniform seed of four digits is created using the
uniform random number
}
