% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{remove_line}
\alias{remove_line}
\title{Remove a Line From a Document}
\usage{
remove_line(pvec_src, pvec_pattern)
}
\arguments{
\item{pvec_src}{vector containing source document}

\item{pvec_pattern}{pattern which identifies line to be removed}
}
\value{
vec_result vector containing source without line to be removed
}
\description{
A line identified by a given pattern is removed from a
document.
}
