% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{remove_section}
\alias{remove_section}
\title{Remove Section From Document}
\usage{
remove_section(pvec_src, pl_tag)
}
\arguments{
\item{pvec_src}{Rmd source vector}

\item{pl_tag}{List with section tags}
}
\value{
vec_result vector with source without section to be removed
}
\description{
A section is removed from a source document where the
section to be removed is indicated by a start and end tag.
}
