% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rtt_log_info}
\alias{rtt_log_info}
\title{Wrapper for log4r info}
\usage{
rtt_log_info(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_caller}{function from which we are called}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r info
}
\examples{
\dontrun{
rtt_logger <- get_rtt_logger()
rtt_log_level(rtt_logger, 'INFO')
rtt_log_info(rtt_logger, 'Examples', 'test message')
}
}
