% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rtt_log_level}
\alias{rtt_log_level}
\title{Wrapper to set the level of a logger}
\usage{
rtt_log_level(plogger, ps_level = c("DEBUG", "INFO", "WARN", "ERROR", "FATAL"))
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_level}{new level of plogger}
}
\description{
Wrapper to set the level of a logger
}
\examples{
\dontrun{
rtt_logger <- get_rtt_logger()
rtt_log_level(rtt_logger, 'INFO')
}
}
