% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cnotes.R
\name{create_cnotes}
\alias{create_cnotes}
\title{Create Course Notes}
\usage{
create_cnotes(
  ps_cn_dir,
  ps_template = "cnotesdown",
  ps_package = "rteachtools",
  pl_repl_data,
  pb_force = FALSE,
  pb_debug = FALSE,
  pobj_rtt_logger = NULL
)
}
\arguments{
\item{ps_cn_dir}{course notes directory}

\item{ps_template}{template for course notes}

\item{ps_package}{package containing course notes template}

\item{pl_repl_data}{list with replacement data}

\item{pb_force}{force to create course notes directory, if it already exists}

\item{pb_debug}{run function in debug mode and write debug info to logfile}

\item{pobj_rtt_logger}{logger object to write loginfo to}
}
\description{
Wrapper function around rmarkdown::draft to create customized course notes.
}
\details{
The function takes care of directory management and replaces placeholders in
templates using data specified in pl_repl_data. The replacement is done using
whisker::whisker.render().
}
\examples{
\dontrun{
# example code
l_repl_data = list(title = "Livestock Breeding and Genomics",
author = "Peter von Rohr",
date = format(Sys.Date(), "\%Y-\%m-\%d"),
url = "https://charlotte-ngs.github.io/lbgfs2023",
github_repo = "https://github.com/charlotte-ngs/lbgfs2023",
book_filename = "lbgfs2023-cn",
book_title = "Lifestock Breeding and Genomics FS 2023",
edit_url = "https://github.com/charlotte-ngs/lbgfs2023/edit/main/\%s")
rteachtools::create_cnotes(ps_cn_dir = "cn_rtt",
                           pl_repl_data = l_repl_data)
}

}
